Initial commit.
Starting out by finding all projects Github considers to be
predominantly written in Rust. Stuck on figuring out how to repeat the
query to get all of the result pages instead of just the first page.
Also stuck on how to store the results and how much of the results
should be stored.