The script works.
Figured out callbacks sufficiently to get the script working. Obtained
all of the search results for Rust repos created after 01-01-2010, and
wrote the results to two files: completed_repos.json and
simple_repos.json.

These are only the repositories that Github considers to be Rust
repositories. Should come up with better search criteria in the future.